requirements.txt
alpaca-py>=0.10.0
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
tavily-python>=0.2.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.15
